Backend/requirements.txt
flask>=3.0
flask-cors>=4.0
pymupdf>=1.24
scikit-learn>=1.4
pandas>=2.2
fpdf2>=2.7
